Remove check from fee quoter
